add all protected route and make the user can only do in their own goal

+++ backend/controller/goalController.js
@@ -1,11 +1,12 @@
 const asyncHandler = require('express-async-handler')
-const goalModel = require('../models/goalModel')
+const goalModel = require('../models/goalModel');
+const userModel = require("../models/userModel")
 
 // @desc Get goals
 // @route GET /api/goals
 // @access Private
 const getGoals = asyncHandler(async (req,res) => {
-    const goals = await goalModel.find()
+    const goals = await goalModel.find({user: req.user.id});
 
         res.status(200).json({
             statusCode: 200,
@@ -22,7 +23,8 @@ const setGoal = asyncHandler(async (req,res) => {
         throw new Error("Please fill the text field")
     }
     const goal = await goalModel.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user.id
     })
 
     res.status(200).json({
@@ -36,12 +38,27 @@ const setGoal = asyncHandler(async (req,res) => {
 // @route PUT /api/goals/:id
 // @access Private
 const updateGoal = asyncHandler(async (req,res) => {
-    const goal = goalModel.findById(req.params.id)
+    const goal = await goalModel.findById(req.params.id);
+
+    const user = await userModel.findById(req.user.id);
+
+    if(!user){
+        res.status(401);
+        throw new Error("User not Found");
+    }
 
     if(!goal){
         res.status(404)
         throw new Error('Goal not found')
     }
+
+
+    // make sure the login user matches the goal user
+    if(goal.user.toString() !== user.id){
+        res.status(401);
+        throw new Error("User not authorized")
+    }
+
 
     const updatedGoal = await goalModel.findByIdAndUpdate(req.params.id,req.body,{
         new: true,
@@ -60,10 +77,21 @@ const updateGoal = asyncHandler(async (req,res) => {
 // @access Private
 const deleteGoal = asyncHandler(async (req,res) => {
     const goal = await goalModel.findById(req.params.id);
+    const user = await userModel.findById(req.user.id);
+
+    if(!user){
+        res.status(401);
+        throw new Error("User not found");
+    }
 
     if(!goal){
         res.status(404)
         throw new Error("Goal not found")
+    }
+
+    if(goal.user.toString() !== user.id){
+        res.status(401);
+        throw new Error("User not authorized")
     }
 
     const deletedGoal = await goalModel.findByIdAndDelete(req.params.id)
